Fix circular module dependency
